chore: Move context decorator into rt module
