+++ py/torch_tensorrt/dynamo/runtime/_PythonTorchTensorRTModule.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import logging
 from contextlib import nullcontext
+from functools import wraps
 from tempfile import tempdir
 from typing import Any, Dict, List, Optional, Sequence, Tuple
 
@@ -21,6 +22,22 @@ from torch_tensorrt.runtime._utils import (
 )
 
 logger = logging.getLogger(__name__)
+
+
+def recreate_context_decorator(method):
+    """
+    A decorator that destroys a context before a method execution and
+    creates it after the method execution within the same class instance.
+    """
+
+    @wraps(method)
+    def wrapper(self, *args, **kwargs):
+        self.reset_context()
+        result = method(self, *args, **kwargs)
+        self.init_context()
+        return result
+
+    return wrapper
 
 
 class PythonTorchTensorRTModule(Module):  # type: ignore[misc]
@@ -126,7 +143,11 @@ class PythonTorchTensorRTModule(Module):  # type: ignore[misc]
     def get_weight_streaming_budget(self):
         return self.engine.weight_streaming_budget_v2
 
+    @recreate_context_decorator
     def set_weight_streaming_budget(self, budget_bytes):
+        return self._set_weight_streaming_budget(budget_bytes)
+
+    def _set_weight_streaming_budget(self, budget_bytes):
         # Disable weight streaming for invalid budget size
         if budget_bytes <= 0:
             budget_bytes = self.get_streamable_weights_size()
@@ -142,7 +163,7 @@ class PythonTorchTensorRTModule(Module):  # type: ignore[misc]
 
     def set_automatic_streaming_budget(self):
         budget_bytes = self.engine.get_weight_streaming_automatic_budget()
-        return self.set_weight_streaming_budget(budget_bytes)
+        return self._set_weight_streaming_budget(budget_bytes)
 
     def setup_engine(self) -> None:
         assert (
