+++ py/torch_tensorrt/dynamo/runtime/_TorchTensorRTModule.py
@@ -4,6 +4,7 @@ import base64
 import copy
 import logging
 import pickle
+from functools import wraps
 from typing import Any, List, Optional, Tuple, Union
 
 import torch
@@ -47,6 +48,22 @@ if ENABLED_FEATURES.torch_tensorrt_runtime:
     SERIALIZED_METADATA_IDX = torch.ops.tensorrt.SERIALIZED_METADATA_IDX()  # 7
     TARGET_PLATFORM_IDX = torch.ops.tensorrt.TARGET_PLATFORM_IDX()  # 8
     SERIALIZATION_LEN = torch.ops.tensorrt.SERIALIZATION_LEN()  # 9
+
+
+def recreate_context_decorator(method):
+    """
+    A decorator that destroys a context before a method execution and
+    creates it after the method execution within the same class instance.
+    """
+
+    @wraps(method)
+    def wrapper(self, *args, **kwargs):
+        self.reset_context()
+        result = method(self, *args, **kwargs)
+        self.init_context()
+        return result
+
+    return wrapper
 
 
 @for_all_methods(needs_torch_tensorrt_runtime)
@@ -181,7 +198,11 @@ class TorchTensorRTModule(torch.nn.Module):  # type: ignore[misc]
     def get_weight_streaming_budget(self):
         return self.engine.weight_streaming_budget_v2
 
+    @recreate_context_decorator
     def set_weight_streaming_budget(self, budget_bytes):
+        return self._set_weight_streaming_budget(budget_bytes)
+
+    def _set_weight_streaming_budget(self, budget_bytes):
         # Disable weight streaming for invalid budget size
         if budget_bytes <= 0:
             budget_bytes = self.get_streamable_weights_size()
@@ -197,7 +218,7 @@ class TorchTensorRTModule(torch.nn.Module):  # type: ignore[misc]
 
     def set_automatic_streaming_budget(self):
         budget_bytes = self.engine.get_weight_streaming_automatic_budget()
-        return self.set_weight_streaming_budget(budget_bytes)
+        return self._set_weight_streaming_budget(budget_bytes)
 
     def setup_engine(self) -> None:
         """
