+++ py/torch_tensorrt/runtime/__init__.py
@@ -7,7 +7,5 @@ from torch_tensorrt.runtime._cudagraphs import (
     get_cudagraphs_mode,
     set_cudagraphs_mode,
 )
-from torch_tensorrt.runtime._weight_streaming import (
-    enable_weight_streaming,
-)
 from torch_tensorrt.runtime._multi_device_safe_mode import set_multi_device_safe_mode
+from torch_tensorrt.runtime._weight_streaming import weight_streaming_context
